+++ tests/Query/QueryTest.php
@@ -9,9 +9,13 @@ use Remorhaz\JSON\Path\Query\Query;
 use Remorhaz\JSON\Path\Query\QueryInterface;
 use Remorhaz\JSON\Path\Query\CapabilitiesInterface;
 use Remorhaz\JSON\Path\Runtime\EvaluatorInterface;
+use Remorhaz\JSON\Path\Runtime\LiteralFactoryInterface;
+use Remorhaz\JSON\Path\Runtime\Matcher\MatcherFactoryInterface;
 use Remorhaz\JSON\Path\Runtime\RuntimeInterface;
+use Remorhaz\JSON\Path\Runtime\ValueListFetcherInterface;
 use Remorhaz\JSON\Path\Value\NodeValueListInterface;
 use Remorhaz\JSON\Path\Value\ValueListInterface;
+use function array_shift;
 
 /**
  * @covers \Remorhaz\JSON\Path\Query\Query
@@ -23,26 +27,38 @@ class QueryTest extends TestCase
     {
         $rootValue = $this->createMock(NodeValueInterface::class);
         $runtime = $this->createMock(RuntimeInterface::class);
+        $valueListFetcher = $this->createMock(ValueListFetcherInterface::class);
         $evaluator = $this->createMock(EvaluatorInterface::class);
+        $literalFactory = $this->createMock(LiteralFactoryInterface::class);
+        $matcherFactory = $this->createMock(MatcherFactoryInterface::class);
+        $runtime
+            ->method('getValueListFetcher')
+            ->willReturn($valueListFetcher);
         $runtime
             ->method('getEvaluator')
             ->willReturn($evaluator);
+        $runtime
+            ->method('getLiteralFactory')
+            ->willReturn($literalFactory);
+        $runtime
+            ->method('getMatcherFactory')
+            ->willReturn($matcherFactory);
 
         $isCallbackCalledWithMatchingArgs = null;
-        $callback = function (
-            NodeValueListInterface $inputArg,
-            RuntimeInterface $runtimeArg,
-            EvaluatorInterface $evaluatorArg
-        ) use (
+        $callback = function () use (
             $rootValue,
-            $runtime,
+            $valueListFetcher,
             $evaluator,
+            $literalFactory,
+            $matcherFactory,
             &$isCallbackCalledWithMatchingArgs
         ): ValueListInterface {
+            $args = func_get_args();
+            /** @var NodeValueListInterface $input */
+            $input = array_shift($args);
             $isCallbackCalledWithMatchingArgs =
-                $inputArg->getValues() === [$rootValue] &&
-                $runtimeArg === $runtime &&
-                $evaluatorArg === $evaluator;
+                $input->getValues() === [$rootValue] &&
+                [$valueListFetcher, $evaluator, $literalFactory, $matcherFactory] === $args;
 
             return $this->createMock(ValueListInterface::class);
         };
@@ -53,7 +69,7 @@ class QueryTest extends TestCase
             $this->createMock(CapabilitiesInterface::class)
         );
 
-        $query($rootValue, $runtime, $evaluator);
+        $query($rootValue, $runtime);
         self::assertTrue($isCallbackCalledWithMatchingArgs);
     }
 
@@ -72,7 +88,6 @@ class QueryTest extends TestCase
         $actualValue = $query(
             $this->createMock(NodeValueInterface::class),
             $this->createMock(RuntimeInterface::class),
-            $this->createMock(EvaluatorInterface::class),
         );
         self::assertSame($values, $actualValue);
     }